+Added menu, difficulties, leaderboard

+++ Igra/src/Source.cpp
@@ -14,7 +14,7 @@ int main(int argc, char* argv[])
 	const int frameDelay = 1000 / FPS;
 
 	Uint32 frameStart;
-	int frameTime;
+	double frameTime;
 
 	game->init("Igra", SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, WindowWidth, WindowHeight, 0);
